pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "goit-pythonweb-hw-08"
version = "0.1.0"
description = "FastAPI contacts management application"
authors = [{name = "Ivan Boichuk", email = "ivan@example.com"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "email-validator>=2.0.0",
    "alembic>=1.12.0",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-dotenv>=1.0.0",
    "slowapi>=0.1.9",
    "redis>=4.0.0",
    "cloudinary>=1.36.0",
    "python-multipart>=0.0.6",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "fastapi-mail>=1.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true